omniparser_service/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.34.0
pydantic-settings~=2.7.0
Pillow~=11.2.1
torch~=2.10.0
easyocr~=1.7.2
torchvision~=0.25.0
supervision==0.18.0
transformers~=4.48.0
ultralytics==8.3.70
numpy==1.26.4
opencv-python-headless~=4.11.0.86
accelerate~=1.12.0
timm~=1.0.24
einops==0.8.0
paddlepaddle~=3.3.0
paddleocr~=2.10.0
pytest~=9.0.2
httpx~=0.28.1
